requirements.txt
gunicorn==20.0.4
Django==3.1.3
line-bot-sdk==1.17.0
gspread==3.7.0
oauth2client==4.1.3
Pillow==8.2.0
beautifulsoup4==4.9.3
apscheduler==3.7.0
numpy==1.19.5
redis==2.10.6
redis-server ==6.0.9